// opencdx-dashboard/src/api/questionnaire/model/questionnaire-item.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { AnfStatementConnector } from './anf-statement-connector';
// May contain unused imports in some cases
// @ts-ignore
import type { AnswerValue } from './answer-value';
// May contain unused imports in some cases
// @ts-ignore
import type { Code } from './code';
// May contain unused imports in some cases
// @ts-ignore
import type { QuestionnaireEnableWhen } from './questionnaire-enable-when';
// May contain unused imports in some cases
// @ts-ignore
import type { QuestionnaireItemAnswerOption } from './questionnaire-item-answer-option';
// May contain unused imports in some cases
// @ts-ignore
import type { QuestionnaireItemExtension } from './questionnaire-item-extension';
// May contain unused imports in some cases
// @ts-ignore
import type { QuestionnaireItemInitial } from './questionnaire-item-initial';

/**
 * 
 * @export
 * @interface QuestionnaireItem
 */
export interface QuestionnaireItem {
    /**
     * 
     * @type {string}
     * @memberof QuestionnaireItem
     */
    'type'?: string;
    /**
     * 
     * @type {string}
     * @memberof QuestionnaireItem
     */
    'linkId'?: string;
    /**
     * 
     * @type {string}
     * @memberof QuestionnaireItem
     */
    'text'?: string;
    /**
     * 
     * @type {Array<QuestionnaireEnableWhen>}
     * @memberof QuestionnaireItem
     */
    'enableWhen'?: Array<QuestionnaireEnableWhen>;
    /**
     * 
     * @type {string}
     * @memberof QuestionnaireItem
     */
    'enableBehavior'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof QuestionnaireItem
     */
    'required'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof QuestionnaireItem
     */
    'repeats'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof QuestionnaireItem
     */
    'readOnly'?: boolean;
    /**
     * 
     * @type {Array<Code>}
     * @memberof QuestionnaireItem
     */
    'code'?: Array<Code>;
    /**
     * 
     * @type {Array<QuestionnaireItemExtension>}
     * @memberof QuestionnaireItem
     */
    'extension'?: Array<QuestionnaireItemExtension>;
    /**
     * 
     * @type {Array<QuestionnaireItemAnswerOption>}
     * @memberof QuestionnaireItem
     */
    'answerOption'?: Array<QuestionnaireItemAnswerOption>;
    /**
     * 
     * @type {Array<QuestionnaireItemInitial>}
     * @memberof QuestionnaireItem
     */
    'initial'?: Array<QuestionnaireItemInitial>;
    /**
     * 
     * @type {Array<AnfStatementConnector>}
     * @memberof QuestionnaireItem
     */
    'anfStatementConnector'?: Array<AnfStatementConnector>;
    /**
     * 
     * @type {Array<AnswerValue>}
     * @memberof QuestionnaireItem
     */
    'answer'?: Array<AnswerValue>;
}