// opencdx-dashboard/src/api/classification/model/rule-set.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface RuleSet
 */
export interface RuleSet {
    /**
     * 
     * @type {string}
     * @memberof RuleSet
     */
    'ruleId'?: string;
    /**
     * 
     * @type {string}
     * @memberof RuleSet
     */
    'type'?: string;
    /**
     * 
     * @type {string}
     * @memberof RuleSet
     */
    'category'?: string;
    /**
     * 
     * @type {string}
     * @memberof RuleSet
     */
    'description'?: string;
    /**
     * 
     * @type {string}
     * @memberof RuleSet
     */
    'name'?: string;
}